+++ src/docs/components/site-header.svelte
@@ -3,6 +3,7 @@
 	import { Discord, GitHub, MobileNav } from '$docs/components/index.js';
 	import { navConfig, siteConfig } from '$docs/config.js';
 	import { cn } from '$docs/utils';
+	import Search from '$routes/(landing-ui)/search.svelte';
 	import Logo from './logo.svelte';
 
 	$: isRoot = $page.url.pathname === '/';
@@ -36,7 +37,7 @@
 				</li>
 			{/each}
 		</ul>
-		<div class="ml-6 flex items-center border-neutral-700 pl-6 md:border-l">
+		<div class="ml-6 flex items-center gap-6 border-neutral-700 pl-6 md:border-l">
 			<a
 				href={siteConfig.links.discord}
 				target="_blank"
@@ -50,11 +51,12 @@
 				href={siteConfig.links.github}
 				target="_blank"
 				rel="noopener noreferrer"
-				class="ml-6 text-neutral-400 transition-colors hover:text-neutral-50"
+				class="text-neutral-400 transition-colors hover:text-neutral-50"
 			>
 				<GitHub class="h-5 w-5" />
 				<span class="sr-only">View the Melt UI GitHub Repository</span>
 			</a>
+			<Search />
 			<MobileNav />
 		</div>
 	</nav>