+++ src/lib/builders/menu/types.ts
@@ -1,6 +1,6 @@
 import type { FloatingConfig } from '$lib/internal/actions/index.js';
 import type { TextDirection } from '$lib/internal/types.js';
-import type { ChangeFn } from '$lib/internal/helpers/index.js';
+import type { ChangeFn, FocusProp } from '$lib/internal/helpers/index.js';
 import type { Writable } from 'svelte/store';
 import type { createMenuBuilder } from './create.js';
 
@@ -99,9 +99,23 @@ export type _CreateMenuProps = {
 	 * @default true
 	 */
 	typeahead?: boolean;
+
+	/**
+	 * Override the default autofocus behavior of the menu
+	 * on close.
+	 */
+	closeFocus?: FocusProp;
+
+	/**
+	 * Optionally prevent focusing the first item in the menu
+	 */
+	disableFocusFirstItem?: boolean;
 };
 
-export type _CreateSubmenuProps = Pick<_CreateMenuProps, 'arrowSize' | 'positioning'> & {
+export type _CreateSubmenuProps = Pick<
+	_CreateMenuProps,
+	'arrowSize' | 'positioning' | 'open' | 'onOpenChange'
+> & {
 	disabled?: boolean;
 };
 
@@ -152,9 +166,10 @@ export type _MenuBuilderOptions = {
 		portal: Writable<string | HTMLElement | undefined | null>;
 		forceVisible: Writable<boolean>;
 		typeahead: Writable<boolean>;
+		closeFocus: Writable<FocusProp | undefined>;
+		disableFocusFirstItem: Writable<boolean>;
 	};
-	disableTriggerRefocus?: boolean;
-	disableFocusFirstItem?: boolean;
+
 	nextFocusable: Writable<HTMLElement | null>;
 	prevFocusable: Writable<HTMLElement | null>;
 	selector: string;