+++ src/lib/shared/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Exports that are shared between multiple builders
+ * can be exported here to prevent conflicts.
+ */
+import type { FocusProp, FocusTarget } from '$lib/internal/helpers/index.js';
+export type { FocusProp, FocusTarget };