+++ src/tests/popover/Popover.spec.ts
@@ -2,49 +2,51 @@ import { render, waitFor } from '@testing-library/svelte';
 import { axe } from 'jest-axe';
 import PopoverTest from './PopoverTest.svelte';
 import userEvent from '@testing-library/user-event';
-import { kbd } from '$lib/internal/helpers/index.js';
+import type { CreatePopoverProps } from '$lib';
+import { testKbd as kbd } from '../utils';
+
+function setup(props: CreatePopoverProps = {}) {
+	const user = userEvent.setup();
+	const returned = render(PopoverTest, props);
+	const trigger = returned.getByTestId('trigger');
+	const content = returned.getByTestId('content');
+	return {
+		user,
+		trigger,
+		content,
+		...returned,
+	};
+}
 
 describe('Popover (Default)', () => {
 	test('No accessibility violations', async () => {
-		const { container } = await render(PopoverTest);
+		const { container } = render(PopoverTest);
 
 		expect(await axe(container)).toHaveNoViolations();
 	});
 
 	test('Opens on click', async () => {
-		const { getByTestId } = await render(PopoverTest);
-		const user = userEvent.setup();
-		const trigger = getByTestId('trigger');
-		const content = getByTestId('content');
+		const { content, trigger, user } = setup();
 
-		await expect(trigger).toBeVisible();
-		await expect(content).not.toBeVisible();
+		expect(content).not.toBeVisible();
 		await user.click(trigger);
 		await waitFor(() => expect(content).toBeVisible());
 	});
 
 	test('Closes on escape', async () => {
-		const { getByTestId } = await render(PopoverTest);
-		const user = userEvent.setup();
-		const trigger = getByTestId('trigger');
-		const content = getByTestId('content');
+		const { content, trigger, user } = setup();
 
-		await expect(trigger).toBeVisible();
-		await expect(content).not.toBeVisible();
+		expect(content).not.toBeVisible();
 		await user.click(trigger);
 		await waitFor(() => expect(content).toBeVisible());
-		await user.keyboard(`{${kbd.ESCAPE}}`);
+		await user.keyboard(kbd.ESCAPE);
 		await waitFor(() => expect(content).not.toBeVisible());
 	});
 
 	test('Closes when click outside content', async () => {
-		const { getByTestId } = await render(PopoverTest);
-		const user = userEvent.setup();
-		const trigger = getByTestId('trigger');
-		const content = getByTestId('content');
+		const { content, trigger, user, getByTestId } = setup();
 
-		await expect(trigger).toBeVisible();
-		await expect(content).not.toBeVisible();
+		expect(content).not.toBeVisible();
 		await user.click(trigger);
 		await waitFor(() => expect(content).toBeVisible());
 		await user.click(content);
@@ -52,5 +54,33 @@ describe('Popover (Default)', () => {
 		const outside = getByTestId('outside');
 		await user.click(outside);
 		await waitFor(() => expect(content).not.toBeVisible());
+	});
+
+	test('Respects `openFocus` prop', async () => {
+		const { content, trigger, user, getByTestId } = setup({
+			openFocus: () => {
+				return document.getElementById('openFocus');
+			},
+		});
+
+		await user.click(trigger);
+		await waitFor(() => expect(content).toBeVisible());
+		const openFocus = getByTestId('openFocus');
+		expect(openFocus).toHaveFocus();
+	});
+
+	test('Respects `closeFocus` prop', async () => {
+		const { content, trigger, user, getByTestId } = setup({
+			closeFocus: () => {
+				return document.getElementById('closeFocus');
+			},
+		});
+
+		await user.click(trigger);
+		await waitFor(() => expect(content).toBeVisible());
+		await user.keyboard(kbd.ESCAPE);
+		await waitFor(() => expect(content).not.toBeVisible());
+		const closeFocus = getByTestId('closeFocus');
+		expect(closeFocus).toHaveFocus();
 	});
 });