+++ svelte.config.js
@@ -1,10 +1,15 @@
-import adapter from '@sveltejs/adapter-vercel';
-// import adapter from '@sveltejs/adapter-static';
+import adapterVercel from '@sveltejs/adapter-vercel';
+import adapterStatic from '@sveltejs/adapter-static';
 import { vitePreprocess } from '@sveltejs/kit/vite';
 import { preprocessMeltUI } from '@melt-ui/pp';
 import { mdsvex } from '@huntabyte/mdsvex';
 import { mdsvexOptions } from './mdsvex.config.js';
 import sequence from 'svelte-sequential-preprocessor';
+
+const IS_PREDEPLOY = process.env.PREDEPLOY === '1';
+const IS_VERCEL = process.env.VERCEL === '1';
+
+const adapter = IS_PREDEPLOY || !IS_VERCEL ? adapterStatic() : adapterVercel({ runtime: 'edge' });
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
@@ -17,9 +22,7 @@ const config = {
 		// adapter-auto only supports some environments, see https://kit.svelte.dev/docs/adapter-auto for a list.
 		// If your environment is not supported or you settled on a specific environment, switch out the adapter.
 		// See https://kit.svelte.dev/docs/adapters for more information about adapters.
-		adapter: adapter({
-			runtime: 'edge',
-		}),
+		adapter,
 		prerender: {
 			handleMissingId: 'ignore',
 		},